Fixing scaleExtent between 0.01 and 100 and initial zoom to scale 1 when zooming in the block view.
Adding zoom rectangle to point zoom event to the complete block view.

Adding reset button to go to the original scale in the block view.

Using the same color for connections (sandybrown) and using the same genome chromosome colors in the block view.

Increasing tickSize to 15 in the block view and changing color scheme to schemeSet2 (using only 8 colors in total) in both views.