Improving dragging functionality:
- Adding transition when the user drags a chromosome, so that the others move back nicely.

- Adding a dictionary of dragging angles having start, end, and total angle for each chromosome.

- Fixing dragging jump when a chromosome is dropped, by keeping track of the last rotation angle and updating the overall genome rotation when chromosomes are not aligned properly.

- Fixing chromosomes collisions by using angles, instead of mouse over event.

- Adding highlight to chromosomes when they are dragged.

Cleaning comments and adding essential documentation to drag functionality.

Decreasing flipping animation time for both the block and the genome view.

Decreasing highlighting opacity for the blocks that are not selected from 0.35 instead to 0.30.

Changing main container to svg instead of div, and having zoom code for genome view in place (commented now).